requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP客戶端
httpx==0.25.2
requests==2.31.0

# 圖片處理
Pillow==10.1.0
opencv-python==4.8.1.78

# 資料處理
pandas==2.1.3
numpy>=1.24.0,<2.0.0

# 環境變數
python-dotenv==1.0.0

# 檔案處理
python-multipart==0.0.6

# 日期處理
python-dateutil==2.8.2

# JSON處理
pydantic==2.5.0
pydantic-settings>=2.0.0

# 日誌
loguru==0.7.2

# 測試
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# 安全檢查
bandit>=1.7.0
safety>=2.3.0

# 開發工具
black>=23.11.0
flake8>=6.1.0
